+++ src/modules/admin/categories/commands/index.ts
@@ -1,0 +1,3 @@
+export * from './create-category'
+export * from './update-category'
+export * from './delete-category'
