+++ src/modules/admin/categories/commands/update-category/index.ts
@@ -1,0 +1,1 @@
+export * from './update-category.resolver'
